Use Flyway to deploy SQL schema to non-prod
Added Gradle tasks to deploy and drop schema in alpha
using Flyway.

Corrected the type of ClaimsEntry's revision_id column.
It should be plain int8, not bigserial.

Make GenerateSqlSchemaCommand use a custom dialect that
converts all varchar type to 'text' and timestamp to
'timestamptz'.

+++ db/src/main/resources/sql/schema/claims_list.sql
@@ -19,7 +19,7 @@ CREATE TABLE `ClaimsList` (
 );
 
 CREATE TABLE `ClaimsEntry` (
-  revision_id BIGSERIAL NOT NULL,
+  revision_id int8 NOT NULL,
   claim_key TEXT NOT NULL,
   domain_label TEXT NOT NULL,
   PRIMARY KEY (revision_id, domain_label),
